fix some eclipse warnings, but in a better (should be more performant) way; trim-whitespace
git-svn-id: https://svn.apache.org/repos/asf/pivot/trunk@1615963 13f79535-47bb-0310-9956-ffa450edef68

+++ wtk/src/org/apache/pivot/wtk/Bounds.java
@@ -77,25 +77,25 @@ public final class Bounds implements Serializable {
         }
 
         if (bounds.containsKey(X_KEY)) {
-            x = (Integer) bounds.get(X_KEY);
+            x = ((Integer) bounds.get(X_KEY)).intValue();
         } else {
             x = 0;
         }
 
         if (bounds.containsKey(Y_KEY)) {
-            y = (Integer) bounds.get(Y_KEY);
+            y = ((Integer) bounds.get(Y_KEY)).intValue();
         } else {
             y = 0;
         }
 
         if (bounds.containsKey(WIDTH_KEY)) {
-            width = (Integer) bounds.get(WIDTH_KEY);
+            width = ((Integer) bounds.get(WIDTH_KEY)).intValue();
         } else {
             width = 0;
         }
 
         if (bounds.containsKey(HEIGHT_KEY)) {
-            height = (Integer) bounds.get(HEIGHT_KEY);
+            height = ((Integer) bounds.get(HEIGHT_KEY)).intValue();
         } else {
             height = 0;
         }
